Extract event utils into own file. Get login configged and up and running. Fix refs to now-extracted funcs left in main events file.
